bubbletea: isolates choicesmodel into its own file
